Add files via upload
•	Designed a Power BI dashboard for Shopping Mall in which having 4000 Customers details. 
(Analyse the Item purchased on the basis of Categories and Season, Compare the Present Purchased amount to Previous Purchased amount, Find the Highest Payment Method used for purchasing, to count the Gender (no. of male & no. of female) on the basis of Categories and Items Purchased, find out the total number of customer ID & total sum of reviews as per customer given, etc.)